Return Mobile option
This change pushes the responsibility on the caller to deal with a
failure to complete the registration. Notice, however, that while it
looks like two separate exit points, it's actually the same code that
has to run in both cases. In other words, this refactoring reduces a DRY
violation.

It's a little odd, however, that an option is being used to return an
'error' as the value, while None indicates success. That's not
idiomatic, but this is easy to fix.